scripts/requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
